pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hozo"
version = "0.1.0"
description = "Wake-on-demand ZFS backup orchestrator"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "hozo contributors"}
]
keywords = ["zfs", "backup", "wake-on-lan", "syncoid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "wakeonlan>=3.1.0",
    "paramiko>=3.4.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "types-PyYAML>=6.0.0",
    "types-paramiko>=3.4.0",
    "httpx>=0.27.0",
]

[project.scripts]
hozo = "hozo.cli:main"
backupd = "backupd.server:run"

[project.urls]
Homepage = "https://github.com/yourusername/hozo"
Repository = "https://github.com/yourusername/hozo"

# ============================================
# Tool Configuration
# ============================================

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".mypy_cache", ".pytest_cache", "venv", ".venv"]

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
packages = ["hozo", "backupd"]
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
